requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Database and Vector Operations
supabase==2.15.2
asyncpg>=0.29.0
sqlalchemy==2.0.23
alembic==1.13.1

# OpenAI Integration (with compatible dependencies)
openai==1.54.3
httpx==0.27.2
distro==1.8.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
python-decouple>=3.8

# HTTP and WebSocket
websockets>=12.0

# Data Processing and Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# Document Processing (optional - will fallback gracefully)
PyPDF2>=3.0.0
pypdf>=4.0.0
pdfplumber>=0.10.0
pymupdf>=1.23.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
# python-magic>=0.4.27  # Commented out due to libmagic system dependency
lxml>=4.9.3
nltk>=3.8.1
tiktoken>=0.5.2
chardet>=5.2.0

# Caching and Performance (optional)
# redis==5.0.1
# aioredis>=2.0.0

# Logging and Monitoring
python-json-logger==2.0.7
structlog==23.2.0

# Utilities
tenacity==8.2.3
numpy==1.24.3
scikit-learn==1.3.2

# Environment and Configuration
python-dotenv==1.0.0

# Development and Testing (optional for production)
pytest>=7.4.0
pytest-asyncio>=0.21.0 